Added pytz zoneinfo files to py2exe's output
pytz has a zoneinfo directory with data, which is in the same directory
as the pytz python files. py2exe, however, only collects python-related
files (such as .pyc). The resulting executable didn't have the time zone
data, resulting in an UnknownTimeZoneError exception, regardless what
time zone was specified in the configuration file.

This commit manually adds the zoneinfo directory to the library.zip
created by py2exe. This appears to be the common way of solving that
problem; see, for example,
http://sourceforge.net/p/py2exe/mailman/message/21872864/
http://sourceforge.net/p/py2exe/mailman/message/21882571/
https://bugs.launchpad.net/openobject-client/+bug/711084
http://stackoverflow.com/questions/9158846/