Simplify CliApp by abolishing check_configuration
CliApp separated read_configuration from check_configuration, but this
was suboptimal and confusing. Although now it interally keeps them
separate (in fact checking is part of reading), it no longer encourages
overriding of check_configuration in subclasses, instead specifying that
the configuration should be checked in overrides of read_configuration.