Ignore grid points with nodata in vaporize
If a GeoTIFF file contained cells with nodata, as is usually the case,
vaporize wasn't ignoring them; instead, it was using the nodata value
(usually a number close to -4e38) in the calculations. Now it properly
converts nodata values to NaN upon reading the file. This was
particularly important for the case of elevation, where the nodata value
was failing the range check.