Fix timeseries flags related tests in Python 3
The timeseries flags is a set of strings; so if there are two flags, say
HEARTS and SPADES, the result when printing the flags can be either
"HEARTS SPADES" or "SPADES HEARTS"; the actual result depends on
circumstances, whereas the related tests were assuming one of the two.
This was always succeding in Python 2, but was sometimes failing in
Python 3.

I think the way I fixed it (accepted both possible outcomes in the
tests) sucks; I think the output should be deterministic, but I'm
leaving proper fixing for when we'll switch to pandas.